+++ HISTORY.md
@@ -1,5 +1,12 @@
 # History
 
+## v0.0.3 - 3 April 2014
+
+* Added support for Wiki printing, along with markdown preprocessor
+* Added tests for wiki URLs
+* Added support for relative image URLs
+* Added test for relative image URLs
+
 ## v0.0.2 - 18 March 2014
 
 * Added URL helper class to translate request URLs to GitHub URLs